# Conflicts:
#	backend/member/src/main/java/com/paymong/member/member/service/MemberService.java

+++ backend/member/src/main/java/com/paymong/member/member/dto/response/FindMemberInfoResDto.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class FindMemberInfoResDto {
     Long mongId;
 
-    Long point;
+    Integer point;
 }
